refactor(list-view-tab): split into DivisionRow + GroupCard subcomponents (sonarjs/no-nested-functions)
The ListViewTab body had two big inline JSX blocks (a draggable
division list inside an IIFE that ran sort+map, and a group
drop-zone .map). Both nested setState callbacks 5 levels deep
(component → IIFE/.map → onChange/onDrop → prev =>), tripping
no-nested-functions for nine call sites.

Lifted into DivisionRow.tsx (per-division list row with the
drag/select handlers) and GroupCard.tsx (per-group drop-zone with
the rename/delete/drop handlers, plus a private findDivisionByKey
helper). The parent now reads sortedDivisions.map() and
subdivisionGroups.map() with no inline closures.
